requirements.txt
langchain==0.1.0
langchain-community==0.0.11
pypdf==3.17.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
streamlit==1.28.0
huggingface-hub==0.19.4
unstructured==0.10.8
markdown==3.4.4
tqdm==4.66.1
sentence-transformers==2.2.2
transformers==4.36.2